managed authorization

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { GetTasksFilterDTO } from './dto/get-task-filter.dto';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { TaskEntity } from './task.entity';
+import { UserEntity } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -13,13 +14,13 @@ export class TasksService {
     @InjectRepository(TaskRepository) private taskRepository: TaskRepository
   ) {}
 
-  getTasks(filterDto: GetTasksFilterDTO): Promise<TaskEntity[]> {
-    return this.taskRepository.getTasks(filterDto)
+  getTasks(filterDto: GetTasksFilterDTO, user: UserEntity): Promise<TaskEntity[]> {
+    return this.taskRepository.getTasks(filterDto, user)
   }
 
 
-  async getTaskById(id: number): Promise<TaskEntity> {
-    const found = await this.taskRepository.findOne({id});
+  async getTaskById(id: number, user: UserEntity): Promise<TaskEntity> {
+    const found = await this.taskRepository.findOne({where: { id, userId: user.id}});
 
     if (!found) {
       throw new NotFoundException('Required task is not found')
@@ -28,20 +29,20 @@ export class TasksService {
     return found; 
   }
 
-  async createTask(createTaskDto: CreateTaskDTO): Promise<TaskEntity> {
-    return this.taskRepository.createTask(createTaskDto);
+  async createTask(createTaskDto: CreateTaskDTO, user: UserEntity): Promise<TaskEntity> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: UserEntity): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id});
 
     if (result.affected === 0 ) {
       new NotFoundException('Could not delete unexisting task');
     }
   }
   
-  async updateStatus(id: number, status: TaskStatus): Promise<TaskEntity> {
-    const task = await this.getTaskById(id);
+  async updateStatus(id: number, status: TaskStatus, user: UserEntity): Promise<TaskEntity> {
+    const task = await this.getTaskById(id, user);
 
     task.status = status;
